WebAppAutomationTests/requirements.txt
allure-pytest==2.8.36
allure-python-commons==2.8.36
Appium-Python-Client==1.0.2
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
enum34==1.1.10
iniconfig==1.1.1
lxml==4.6.2
namedlist==1.8
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-rerunfailures==9.1.1
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.3
